Redirect loop with authentication success but access denied at default identity provider closes #17441
