+++ themes/src/main/resources/theme/base/account/messages/messages_en.properties
@@ -214,6 +214,7 @@ identityProviderRemovedMessage=Identity provider removed successfully.
 identityProviderAlreadyLinkedMessage=Federated identity returned by {0} is already linked to another user.
 staleCodeAccountMessage=The page expired. Please try one more time.
 consentDenied=Consent denied.
+access-denied-when-idp-auth=Access denied when authenticating with {0}
 
 accountDisabledMessage=Account is disabled, contact your administrator.
